agentserver/pyproject.toml
[tool.poetry]
name = "agentserver"
version = "0.1.0"
description = ""
authors = ["p-oong <lhp7505@naver.com>"]
license = "MIT"
readme = "README.md"
package-mode = false  

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.17.0"
typing_extensions = "^4.5.0"
python-dotenv = "^1.0.0"
requests = "^2.28.0"
langchain = "0.3.15"
langchain-openai = ">=0.1.26"
langchain-core = ">=0.3.31,<0.4.0"
feedparser = "^6.0.10"
bs4 = "^0.0.1"
mojito2 = "^0.1.6"
pandas = "^2.2.3"
lxml = "^5.3.0"
yfinance = "^0.2.52"
fastapi = "^0.115.8"
sentence-transformers = "^3.4.1"
pydantic = "^2.10.6"
logging = "^0.4.9.6"
elasticsearch = "^8.17.1"
langchain-community = "0.3.15"
uvicorn = "^0.34.0"
sqlmodel = "^0.0.22"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
